refactor + add task progress handler to incremental backup and replication
